Migrate servlets to Spring REST + Java 17
Replace legacy servlet/JSP MVC with Spring REST controllers (add AuthController, BookingController, MovieController, UserController, etc.) and remove many old servlet and JSP/webapp files. Update JWT handling in JwtAuthFilter to use userId as the authentication principal. Update project tooling/config for Java 17 (Eclipse .classpath and .settings changes, enable annotation processing and null analysis) and add spring-boot-starter-actuator to the pom. These changes move the backend toward a Spring Boot REST API, improve JWT principal semantics, and modernize build/runtime configuration.

+++ cinosphere/src/main/java/com/cinosphere/dto/BookingResponse.java
@@ -1,5 +1,6 @@
 package com.cinosphere.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
@@ -14,7 +15,7 @@ public class BookingResponse {
     private LocalDate bookingDate;
     private LocalTime bookingTime;
     private String bookingStatus;
-    private double totalAmount;
+    private BigDecimal totalAmount;
     private int loyaltyPointsEarned;
     private String paymentMethod;
 
@@ -59,8 +60,8 @@ public class BookingResponse {
     public String getBookingStatus()           { return bookingStatus; }
     public void setBookingStatus(String v)     { this.bookingStatus = v; }
 
-    public double getTotalAmount()             { return totalAmount; }
-    public void setTotalAmount(double v)       { this.totalAmount = v; }
+    public BigDecimal getTotalAmount()             { return totalAmount; }
+    public void setTotalAmount(BigDecimal v)       { this.totalAmount = v; }
 
     public int getLoyaltyPointsEarned()        { return loyaltyPointsEarned; }
     public void setLoyaltyPointsEarned(int v)  { this.loyaltyPointsEarned = v; }
