+++ cinosphere/src/main/java/com/cinosphere/dto/UpdateBasePriceRequest.java
@@ -1,5 +1,7 @@
 package com.cinosphere.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotNull;
 
@@ -10,8 +12,8 @@ public class UpdateBasePriceRequest {
 
     @NotNull(message = "Base price is required")
     @DecimalMin(value = "0.01", message = "Base price must be greater than 0")
-    private Double basePrice;
+    private BigDecimal basePrice;
 
-    public Double getBasePrice()        { return basePrice; }
-    public void setBasePrice(Double v)  { this.basePrice = v; }
+    public BigDecimal getBasePrice()        { return basePrice; }
+    public void setBasePrice(BigDecimal v)  { this.basePrice = v; }
 }
